moved helper function that converts data to a separate folder for modularity
